src/requirements/base.txt
python-dotenv==1.0.0
redis==4.5.5
fastapi==0.95.2
orjson==3.8.12
uvicorn==0.22.0
aiohttp==3.8.4
uvloop==0.17.0 ; sys_platform != "win32" and implementation_name == "cpython"
psycopg2-binary==2.9.6
python-multipart==0.0.6
sqlalchemy==2.0.15
passlib==1.7.4
pydantic[email]
PyJWT==2.7.0
alembic==1.11.1